+++ docs/README.md
@@ -22,7 +22,7 @@
   * [expression.md](expression.md) - Loading microarray gene expression data.
   * *under development* - Loading readcounts and normalized readcounts data for mRNAseq and miRNAseq.
 * Copy Number Variation data
-  * [acgh.md](acgh.md) - Loading CNV data from Array CGH (comparative genomic hybridisation), SNP Array, DNA-Seq, etc.
+  * [cnv.md](cnv.md) - Loading CNV data from Array CGH (comparative genomic hybridisation), SNP Array, DNA-Seq, etc.
 * Small Genomic Variants
   * *not yet implemented* - Loading small genomic variants (SNP, indel in VCF format) from RNAseq or DNAseq.
 * Proteomics data
