--- src/main/groovy/org/transmartproject/batch/highdim/acgh/data/AcghDataJobSpecification.groovy
@@ -1,24 +1,0 @@
-package org.transmartproject.batch.highdim.acgh.data
-
-import groovy.transform.TypeChecked
-import org.transmartproject.batch.highdim.jobparams.StandardAssayParametersModule
-import org.transmartproject.batch.highdim.jobparams.StandardHighDimDataParametersModule
-import org.transmartproject.batch.startup.ExternalJobParametersModule
-import org.transmartproject.batch.startup.JobSpecification
-import org.transmartproject.batch.startup.StudyJobParametersModule
-
-/**
- * External parameters for RNASeq data.
- */
-@TypeChecked
-class AcghDataJobSpecification implements JobSpecification {
-
-    final List<? extends ExternalJobParametersModule> jobParametersModules = [
-            new StudyJobParametersModule(),
-            new StandardAssayParametersModule(),
-            new StandardHighDimDataParametersModule(),
-            new AcghDataParametersModule()
-    ]
-
-    final Class<?> jobPath = AcghDataJobConfig
-}
