+++ src/main/groovy/org/transmartproject/batch/startup/JobStartupDetails.groovy
@@ -10,8 +10,8 @@ import org.springframework.batch.core.JobParameters
 import org.transmartproject.batch.backout.BackoutJobSpecification
 import org.transmartproject.batch.clinical.ClinicalJobSpecification
 import org.transmartproject.batch.gwas.GwasJobSpecification
-import org.transmartproject.batch.highdim.acgh.data.AcghDataJobSpecification
-import org.transmartproject.batch.highdim.acgh.platform.AcghAnnotationJobSpecification
+import org.transmartproject.batch.highdim.cnv.data.CnvDataJobSpecification
+import org.transmartproject.batch.highdim.cnv.platform.CnvAnnotationJobSpecification
 import org.transmartproject.batch.highdim.metabolomics.data.MetabolomicsDataJobSpecification
 import org.transmartproject.batch.highdim.metabolomics.platform.MetabolomicsAnnotationJobSpecification
 import org.transmartproject.batch.highdim.mrna.data.MrnaDataJobSpecification
@@ -53,8 +53,8 @@ final class JobStartupDetails {
             'backout'                : BackoutJobSpecification,
             'rnaseq_annotation'      : RnaSeqAnnotationJobSpecification,
             'rnaseq'                 : RnaSeqDataJobSpecification,
-            'acgh_annotation'        : AcghAnnotationJobSpecification,
-            'acgh'                   : AcghDataJobSpecification,
+            'cnv_annotation'         : CnvAnnotationJobSpecification,
+            'cnv'                    : CnvDataJobSpecification,
     ]
     public static final String STUDY_PARAMS_FILE_NAME = 'study' + PARAMS_FILE_EXTENSION
     public static final String PARAMS_FILE_EXTENSION = '.params'
