+++ src/test/groovy/org/transmartproject/batch/startup/JobStartupDetailsTests.groovy
@@ -5,7 +5,7 @@ import org.junit.Test
 import org.junit.rules.ExpectedException
 import org.junit.rules.TemporaryFolder
 import org.transmartproject.batch.clinical.ClinicalDataLoadJobConfiguration
-import org.transmartproject.batch.highdim.acgh.data.AcghDataJobConfig
+import org.transmartproject.batch.highdim.cnv.data.CnvDataJobConfig
 import org.transmartproject.batch.highdim.mrna.data.MrnaDataJobConfig
 import org.transmartproject.batch.highdim.mrna.platform.MrnaPlatformJobConfig
 
@@ -304,32 +304,32 @@ class JobStartupDetailsTests {
     }
 
     @Test
-    void testAcghProbIsNotOneDefaultIsError() {
+    void testCnvProbIsNotOneDefaultIsError() {
         def dataFileName = 'empty_data_file.tsv'
         def ssmFileName = 'empty_ssm_file.tsv'
 
         createTmpFile dataFileName
         createTmpFile ssmFileName
-        Path paramsFile = createTmpFile 'acgh.params', """
+        Path paramsFile = createTmpFile 'cnv.params', """
             STUDY_ID=test_study
             DATA_FILE=${dataFileName}
             MAP_FILENAME=${ssmFileName}
         """
 
         JobStartupDetails startupDetails = JobStartupDetails.fromFile(paramsFile)
-        assertThat startupDetails, hasProperty('jobPath', equalTo(AcghDataJobConfig))
+        assertThat startupDetails, hasProperty('jobPath', equalTo(CnvDataJobConfig))
         assertThat startupDetails['PROB_IS_NOT_1'], equalTo('ERROR')
     }
 
     @Test
-    void testAcghProbIsNotOne() {
+    void testCnvProbIsNotOne() {
         def dataFileName = 'empty_data_file.tsv'
         def ssmFileName = 'empty_ssm_file.tsv'
         def providedArgument = 'WARN'
 
         createTmpFile dataFileName
         createTmpFile ssmFileName
-        Path paramsFile = createTmpFile 'acgh.params', """
+        Path paramsFile = createTmpFile 'cnv.params', """
             STUDY_ID=test_study
             DATA_FILE=${dataFileName}
             MAP_FILENAME=${ssmFileName}
@@ -341,14 +341,14 @@ class JobStartupDetailsTests {
     }
 
     @Test(expected = InvalidParametersFileException)
-    void testAcghProbIsNotOneExpectedError() {
+    void testCnvProbIsNotOneExpectedError() {
         def dataFileName = 'empty_data_file.tsv'
         def ssmFileName = 'empty_ssm_file.tsv'
         def providedArgument = 'FAIL!'
 
         createTmpFile dataFileName
         createTmpFile ssmFileName
-        Path paramsFile = createTmpFile 'acgh.params', """
+        Path paramsFile = createTmpFile 'cnv.params', """
             STUDY_ID=test_study
             DATA_FILE=${dataFileName}
             MAP_FILENAME=${ssmFileName}
