--- studies/ACGH_ANNOT/acgh_annotation/chromosomal_regions.tsv
@@ -1,5 +1,0 @@
-GPL_ID	REGION_NAME	CHROMOSOME	START_BP	END_BP	NUM_PROBES	CYTOBAND	GENE_SYMBOL	GENE_ID	ORGANISM
-ACGH_ANNOT	_WASH7P	1	2	3	4	CYT	WASH7P	567	Homo Sapiens
-ACGH_ANNOT	FAM138F						FAM138F		Homo Sapiens
-ACGH_ANNOT	LOC100132287						LOC100132287		Homo Sapiens
-ACGH_ANNOT	OR4F3						OR4F3		Homo Sapiens
