--- studies/CLUC/acgh.params
@@ -1,4 +1,0 @@
-DATA_FILE_PREFIX=acgh_data.tsv
-DATA_TYPE=R
-LOG_BASE=2
-MAP_FILENAME=acgh_subject_sample_mapping.tsv
