--- studies/CLUC/acgh/acgh_data.tsv
@@ -1,5 +1,0 @@
-region_name	ACGH.CACO2.flag	ACGH.COLO205.flag	ACGH.DLD1.flag	ACGH.CACO2.chip	ACGH.COLO205.chip	ACGH.DLD1.chip	ACGH.CACO2.segmented	ACGH.COLO205.segmented	ACGH.DLD1.segmented	ACGH.CACO2.probhomloss	ACGH.COLO205.probhomloss	ACGH.DLD1.probhomloss	ACGH.CACO2.probloss	ACGH.COLO205.probloss	ACGH.DLD1.probloss	ACGH.CACO2.probnorm	ACGH.COLO205.probnorm	ACGH.DLD1.probnorm	ACGH.CACO2.probgain	ACGH.COLO205.probgain	ACGH.DLD1.probgain	ACGH.CACO2.probamp	ACGH.COLO205.probamp	ACGH.DLD1.probamp
-FAM138F	0	-2	-1	-0.1335	-0.357	-0.143	-0.1335	-0.357	-0.144	0	0.908	0.002	0.092	0	0.880	0.908	0.092	0.112	0	0	0.005	0	0	0.001
-LOC100132287	1	2	0	-0.143	-0.143	0.1345	-0.143	-0.143	0.1345	0	0	0	0.112	0.112	0	0	0	0.763	0.888	0.333	0.237	0	0.555	0
-OR4F3	1	0	0		0.0915	-0.038		0.0915	-0.038		0	0		0	0.038		0.763	0.962		0.237	0		0	0
-_WASH7P	0	0	0	0.666	-0.49	-0.238	0.666	-0.49	-0.238	0	0	0	0	0	0.092	0.711	0.711	0.908	0.289	0.289	0	0	0	0
