--- studies/CLUC/acgh/acgh_subject_sample_mapping.tsv
@@ -1,4 +1,0 @@
-trial_name	site_id	subject_id	sample_cd	platform	sample_type	tissue_type	time_point	cat_cd	src_cd
-CLUC		CACO2	ACGH.CACO2	ACGH_ANNOT	cgh	Colon	t1	"Test+Acgh+SAMPLETYPE+data"	STD
-CLUC		COLO205	ACGH.COLO205	ACGH_ANNOT	cgh	Colon	t1	"Test+Acgh+SAMPLETYPE+data"	STD
-CLUC		DLD1	ACGH.DLD1	ACGH_ANNOT	cgh	Colon	t1	"Test+Acgh+SAMPLETYPE+data"	STD
