+++ studies/CLUC/cnv.params
@@ -1,0 +1,4 @@
+DATA_FILE_PREFIX=cnv_data.tsv
+DATA_TYPE=R
+LOG_BASE=2
+MAP_FILENAME=cnv_subject_sample_mapping.tsv
