+++ studies/CLUC/cnv/cnv_data.tsv
@@ -1,0 +1,5 @@
+region_name	CNV.CACO2.flag	CNV.COLO205.flag	CNV.DLD1.flag	CNV.CACO2.chip	CNV.COLO205.chip	CNV.DLD1.chip	CNV.CACO2.segmented	CNV.COLO205.segmented	CNV.DLD1.segmented	CNV.CACO2.probhomloss	CNV.COLO205.probhomloss	CNV.DLD1.probhomloss	CNV.CACO2.probloss	CNV.COLO205.probloss	CNV.DLD1.probloss	CNV.CACO2.probnorm	CNV.COLO205.probnorm	CNV.DLD1.probnorm	CNV.CACO2.probgain	CNV.COLO205.probgain	CNV.DLD1.probgain	CNV.CACO2.probamp	CNV.COLO205.probamp	CNV.DLD1.probamp
+FAM138F	0	-2	-1	-0.1335	-0.357	-0.143	-0.1335	-0.357	-0.144	0	0.908	0.002	0.092	0	0.880	0.908	0.092	0.112	0	0	0.005	0	0	0.001
+LOC100132287	1	2	0	-0.143	-0.143	0.1345	-0.143	-0.143	0.1345	0	0	0	0.112	0.112	0	0	0	0.763	0.888	0.333	0.237	0	0.555	0
+OR4F3	1	0	0		0.0915	-0.038		0.0915	-0.038		0	0		0	0.038		0.763	0.962		0.237	0		0	0
+_WASH7P	0	0	0	0.666	-0.49	-0.238	0.666	-0.49	-0.238	0	0	0	0	0	0.092	0.711	0.711	0.908	0.289	0.289	0	0	0	0
