--- studies/SKIP_UNMAPPED_DATA/acgh.params
@@ -1,3 +1,0 @@
-DATA_FILE_PREFIX=acgh_data.tsv
-MAP_FILENAME=acgh_subject_sample_mapping.tsv
-SKIP_UNMAPPED_DATA=Y
