--- studies/SKIP_UNMAPPED_DATA/acgh_data.tsv
@@ -1,5 +1,0 @@
-region_name	ACGH.S11.flag	ACGH.S21.flag
-FAM138F	0	2
-LOC100132287	1	2
-OR4F3	1	0
-_WASH7P	0	1
