+++ studies/SKIP_UNMAPPED_DATA/cnv.params
@@ -1,0 +1,3 @@
+DATA_FILE_PREFIX=cnv_data.tsv
+MAP_FILENAME=cnv_subject_sample_mapping.tsv
+SKIP_UNMAPPED_DATA=Y
