+++ studies/SKIP_UNMAPPED_DATA/cnv_data.tsv
@@ -1,0 +1,5 @@
+region_name	CNV.S11.flag	CNV.S21.flag
+FAM138F	0	2
+LOC100132287	1	2
+OR4F3	1	0
+_WASH7P	0	1
