+++ studies/SKIP_UNMAPPED_DATA/cnv_subject_sample_mapping.tsv
@@ -1,0 +1,2 @@
+trial_name	site_id	subject_id	sample_cd	platform	sample_type	tissue_type	time_point	cat_cd	src_cd
+SKIP_UNMAPPED_DATA		S2	CNV.S21	CNV_ANNOT	cnv	Colon	tp1	"Test+Cnv+SAMPLETYPE+data"	STD
